backend/requirements.txt
uvicorn==0.23.2
sqlalchemy==2.0.22
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
apscheduler==3.10.4
pillow==10.1.0
pytest==7.4.3
httpx==0.25.0
pytest-asyncio==0.21.1
fastapi==0.103.2
pydantic==2.3.0
typing-extensions==4.7.1
aiogram==3.1.1
requests
